Integrate auth with todos

+++ routers/todos.py
@@ -4,6 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException, status, Path
 from models import Todos
 from database import engine, SessionLocal
 from pydantic import BaseModel, Field
+from .auth import get_current_user
 
 router = APIRouter(
     prefix="/todo",
@@ -19,6 +20,7 @@ def get_db():
 
 # Depends - it is a dependency injection
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -30,51 +32,65 @@ class TodoRequest(BaseModel):
 
 
 @router.post('/', status_code=status.HTTP_201_CREATED)
-def create_todo(todo: TodoRequest, db: db_dependency):
-    try:
-        todo_model = Todos(**todo.dict())
-        db.add(todo_model)
-        db.commit()
-        return {"message": "Todo created successfully"}
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
+def create_todo(todo: TodoRequest, user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = Todos(**todo.dict(), owner_id=user.get("id"))
+    db.add(todo_model)
+    db.commit()
+    return {"message": "Todo created successfully"}
+
+@router.get('/', status_code=status.HTTP_200_OK)
+def read_all_todos(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    if user.get("role") != "admin":
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todos = db.query(Todos).all()
+    return todos
 
 @router.get('/{todo_id}', status_code=status.HTTP_200_OK)
-def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get("id")).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
 
 @router.put('/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-def update_todo(todo: TodoRequest, db: db_dependency, todo_id: int = Path(gt=0)):
-    try:
-        todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-        if todo_model is None:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
-        todo_model.title = todo.title
-        todo_model.description = todo.description
-        todo_model.priority = todo.priority
-        todo_model.complete = todo.complete
-        db.add(todo_model)
-        db.commit()
-        return {"message": "Todo updated successfully"}
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
+def update_todo(user: user_dependency, todo: TodoRequest, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get("id")).first()
+    if todo_model is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
+    
+    todo_model.title = todo.title
+    todo_model.description = todo.description
+    todo_model.priority = todo.priority
+    todo_model.complete = todo.complete
+    
+    db.add(todo_model)
+    db.commit()
+    
+    return {"message": "Todo updated successfully"}
         
 
 @router.delete('/{todo_id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    try:
-        todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-        if todo_model is None:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
-        db.query(Todos).filter(Todos.id == todo_id).delete()
-        db.commit()
-        return {"message": "Todo deleted successfully"}
-    except Exception as e:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
+def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get("id")).first()
+    if todo_model is None:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Todo not found")
+    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.commit()
+    return {"message": "Todo deleted successfully"}
